WC Blocks: Moved `php` files to `source` for logical consistency.
To follow class standard, it would be more logical to move the `php` source files for components in the same directory as their JavaScript counterparts.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8860 74240141-8908-4e6f-9713-ba540dce6ec7
