+++ public_html/wp-content/mu-plugins/blocks/source/blocks.js
@@ -7,6 +7,7 @@ const { registerBlockType } = wp.blocks;
  * Internal dependencies
  */
 import './_z-index.scss';           // Have z-index values, similar to https://github.com/WordPress/gutenberg/blob/master/assets/stylesheets/_z-index.scss
+import './styles.scss';               // Common styles for WordCamp Blocks.
 import { BLOCKS } from './blocks/'; // Trailing slash required to differentiate the folder from the file.
 
 BLOCKS.forEach( ( { NAME, SETTINGS } ) => {
