+++ public_html/wp-content/mu-plugins/blocks/source/blocks/sessions/view.php
