+++ public_html/wp-content/mu-plugins/blocks/source/blocks/sponsors/view.php
