+++ public_html/wp-content/mu-plugins/blocks/source/components/block-content/style.scss
@@ -6,7 +6,3 @@
 		float: none;
 	}
 }
-
-.wordcamp-item-meta {
-	margin: 0 0 1.5em;
-}
