--- public_html/wp-content/mu-plugins/blocks/source/components/block-controls/style.scss
@@ -1,4 +1,0 @@
-.wordcamp-block-edit-mode-option {
-	flex-basis: 100%;
-	padding: 0.5em;
-}
