+++ public_html/wp-content/mu-plugins/blocks/source/styles.scss
@@ -1,0 +1,8 @@
+.wordcamp__edit-mode-option {
+	flex-basis: 100%;
+	padding: 0.5em;
+}
+
+.wordcamp__post-meta {
+	margin: 0 0 1.5em;
+}
